chore: update with feedback, make toolbar a component
